requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# ML & Data Science (using compatible versions for Windows)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Data Sources
yfinance>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# NLP (lightweight versions)
textblob>=0.17.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Optional ML packages (install separately if needed)
# xgboost>=2.0.0
# shap>=0.43.0
# transformers>=4.36.0